+++ internal/tgbot/bot_outside.go
@@ -2,8 +2,8 @@ package tgbot
 
 import tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
 
-func (b *TgBot) SendText(chatID int, message string) {
-	msg := tgbotapi.NewMessage(int64(chatID), message)
+func (b *TgBot) SendText(chatID int64, message string) {
+	msg := tgbotapi.NewMessage(chatID, message)
 
 	if _, err := b.Bot.Send(msg); err != nil {
 		b.log.WithError(err).Error("send message")
